fix(macros): reject non-usize scalar planning variables
Reject #[planning_variable] fields that are not Option<usize> instead of silently emitting metadata without runtime scalar accessors.

The adjacent audit found the same silent skip behavior in both planning_entity derive expansion and planning_model metadata parsing. That allowed Option<u32> or Option<i64> scalar variables to compile while producing no generated scalar runtime slot, which surfaced as zero moves.

This change makes the contract a compile-time error, adds a trybuild regression for Option<u32>, updates stale facade fixtures that taught Option<i64> planning variables, and documents that scalar variables store candidate indexes while external IDs belong on facts or entities.

+++ crates/solverforge/tests/constraint_accessors/domain/shift.rs
@@ -6,5 +6,5 @@ pub struct Shift {
     pub id: i64,
 
     #[planning_variable]
-    pub employee: Option<i64>,
+    pub employee: Option<usize>,
 }
